@@ -41,6 +41,78 @@ li {
 }
 `;
 
+export const css2 = `.alert {
+  position: relative;
+  padding: 1.6rem 4.6rem;
+  margin-bottom: 1.6rem;
+  border: 1px solid #c53030;
+  color: #fff;
+  border-radius: 0.2rem;
+  width: 100%;
+}
+
+.logo {
+  margin-bottom: 1.6rem;
+  background: url('logo.svg') no-repeat;
+  display: flex;
+  justify-content: center;
+}
+
+.button {
+  background: #81e6d9;
+  padding: 1.6rem 4.6rem;
+  letter-spacing: 0.03rem;
+  border-radius: 0.2rem;
+}
+
+.button:hover {
+  background: #2c7a7b;
+}
+
+@media (min-width: 640px) {
+  .button {
+    padding: 0.5rem 1rem;
+    width: 100%;
+  } 
+}
+
+@media (min-width: 1280px) {
+  .button {
+    padding: 3rem 7rem;
+    margin-bottom: 2.4rem;
+  } 
+}
+
+.username {
+  color: #718096;
+  border-color: #bee3f8;
+}
+
+.username:focus {
+  border-color: #3182ce;
+}
+
+.username::placeholder {
+  color: #cbd5e0;
+}
+
+@media (min-width: 1280px) {
+  .username {
+    width: 50%;
+  } 
+}
+
+.footer {
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  flex-direction: row-reverse;
+  padding: 2.4rem 3rem;
+  border-top: 1px solid #fff5f5;
+}
+`;
+
 export const javascript = `const container = css({
   flex: 1,
   padding: 10,
@@ -333,17 +405,15 @@ export const ShowStoredUser: React.FC<Props> = (props) => {
 
 import { useState, useEffect } from 'react';
 
-export const CounterExample = () => {
+export const CounterExample: React.FC<{}> = () => {
   const [count, setCount] = useState(0);
-
-  useEffect(() => {
-    document.title = \`You clicked $\{count\} times\`;
-  });
+  
+  const handleClick = () => setCount(count + 1)
 
   return (
     <div>
       <p>You clicked {count} times</p>
-      <button onClick={() => setCount(count + 1)}>
+      <button onClick={handleClick}>
         Click me
       </button>
     </div>
