@@ -1,11 +1,20 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { Config } from "ts-json-schema-generator/dist/src/Config";
-import tempy from "tempy";
 import * as tsj from "ts-json-schema-generator";
+import os from "os";
+import crypto from "crypto";
+import path from "path";
+import fs from "fs";
+
+const tmpDir = os.tmpdir?.();
 
 export default (req: NextApiRequest, res: NextApiResponse) => {
+  const filePath =
+    path.join(tmpDir, crypto.randomBytes(16).toString("hex")) + ".ts";
   try {
-    const filePath = tempy.writeSync(req.body, { extension: "ts" });
+    fs.writeFileSync(filePath, req.body, {
+      encoding: "utf-8"
+    });
     const config: Config = {
       path: filePath,
       expose: "all",
@@ -18,4 +27,5 @@ export default (req: NextApiRequest, res: NextApiResponse) => {
   } catch (e) {
     res.status(500).send(e.message);
   }
+  fs.unlinkSync(filePath);
 };
